requirements.txt
keyboard>=0.13.4
python-daemon>=2.2.3
